+++ src/Simplify.Web.Tests/StaticFiles/Handlers/NewFileHandlerTests.cs
@@ -1,10 +1,10 @@
 using System;
+using System.IO;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Moq;
 using NUnit.Framework;
 using Simplify.System;
-using Simplify.Web.Http.ResponseWriting;
 using Simplify.Web.StaticFiles.Context;
 using Simplify.Web.StaticFiles.Handlers;
 using Simplify.Web.StaticFiles.IO;
@@ -17,15 +17,13 @@ public class NewFileHandlerTests
 {
 	private NewFileHandler _handler = null!;
 
-	private Mock<IResponseWriter> _responseWriter = null!;
 	private Mock<IStaticFile> _staticFile = null!;
 
 	[SetUp]
 	public void Initialize()
 	{
-		_responseWriter = new Mock<IResponseWriter>();
 		_staticFile = new Mock<IStaticFile>();
-		_handler = new NewFileHandler(_responseWriter.Object, _staticFile.Object);
+		_handler = new NewFileHandler(_staticFile.Object);
 	}
 
 	[Test]
@@ -45,7 +43,6 @@ public class NewFileHandlerTests
 	public void CanHandle_CantBeCached_True()
 	{
 		// Arrange
-
 		var context = Mock.Of<IStaticFileProcessingContext>(x => !x.CanBeCached);
 
 		// Act
@@ -56,13 +53,12 @@ public class NewFileHandlerTests
 	}
 
 	[Test]
-	public async Task ExecuteAsync_NewFile_FileSendToClientAndRespectiveResponsePropertiesAreSet()
+	public async Task ExecuteAsync_NewFile_FileStreamedToResponseBody()
 	{
 		// Arrange
 
 		var filePath = "Foo.txt";
 		var lastModificationTime = new DateTime(2013, 4, 5, 0, 0, 0, DateTimeKind.Utc);
-		var data = new byte[1] { 255 };
 
 		var context = Mock.Of<IStaticFileProcessingContext>(x =>
 			x.RelativeFilePath == filePath &&
@@ -70,9 +66,14 @@ public class NewFileHandlerTests
 
 		TimeProvider.Current = Mock.Of<ITimeProvider>(x => x.Now == new DateTime(2013, 1, 1, 0, 0, 0, DateTimeKind.Utc));
 
-		_staticFile.Setup(x => x.GetDataAsync(It.Is<string>(s => s == filePath))).Returns(Task.FromResult(data));
+		var bodyStream = new MemoryStream();
+		var response = Mock.Of<HttpResponse>(x =>
+			x.Headers == new HeaderDictionary() &&
+			x.Body == bodyStream);
 
-		var response = Mock.Of<HttpResponse>(x => x.Headers == new HeaderDictionary());
+		_staticFile.Setup(x => x.CopyToAsync(It.Is<Stream>(s => s == bodyStream), It.Is<string>(s => s == filePath)))
+			.Returns(Task.CompletedTask)
+			.Verifiable();
 
 		// Act
 		await _handler.ExecuteAsync(context, response);
@@ -83,6 +84,6 @@ public class NewFileHandlerTests
 		Assert.That(response.Headers["Last-Modified"], Is.EqualTo(lastModificationTime.ToString("r")));
 		Assert.That(response.Headers["Expires"], Is.EqualTo(new DateTimeOffset(new DateTime(2014, 1, 1, 0, 0, 0, DateTimeKind.Utc)).ToString("R")));
 
-		_responseWriter.Verify(x => x.WriteAsync(It.Is<HttpResponse>(r => r == response), It.Is<byte[]>(b => b == data)));
+		_staticFile.Verify();
 	}
 }