+++ src/Simplify.Web/StaticFiles/Handlers/NewFileHandler.cs
@@ -1,6 +1,5 @@
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
-using Simplify.Web.Http.ResponseWriting;
 using Simplify.Web.StaticFiles.Context;
 using Simplify.Web.StaticFiles.IO;
 
@@ -10,7 +9,7 @@ namespace Simplify.Web.StaticFiles.Handlers;
 /// Provides the new file handler.
 /// </summary>
 /// <seealso cref="IStaticFileRequestHandler" />
-public class NewFileHandler(IResponseWriter responseWriter, IStaticFile staticFile) : IStaticFileRequestHandler
+public class NewFileHandler(IStaticFile staticFile) : IStaticFileRequestHandler
 {
 	/// <summary>
 	/// Determines whether this handler can handle the file requested.
@@ -30,6 +29,6 @@ public class NewFileHandler(IResponseWriter responseWriter, IStaticFile staticFi
 	{
 		response.SetNewReturningFileAttributes(context);
 
-		await responseWriter.WriteAsync(response, await staticFile.GetDataAsync(context.RelativeFilePath));
+		await staticFile.CopyToAsync(response.Body, context.RelativeFilePath);
 	}
 }